# speed up read.beast