## Description

read.beast is very slow. This PR speeds up the code by using perl compatible regexes. In a benchmark this is about ~79% faster:

```
label           min     median    mem_alloc  total_time  n_itr
original (OLD)  17.7s      18s    60.3MB      3.01m    10
pcre (NEW)      3.71s    3.77s    60.3MB     37.71s    10
```

I've also added a more complete test case to make sure the beast trees are loaded correctly, and the relevant tree annotations are loaded properly.


## Related Issue

#116 & #117 are merged here as I was finding the test warnings annoying.