@SimonGreenhill 2023-11-23
Checking whether the treefile is translated or not is expensive, so ff319a2 also only does this once and gain another 4% decrease in run time.