@SimonGreenhill 2023-11-23
This is now a 82% decrease in speed. 

```
  label          min   median mem_alloc total_time n_itr
  <chr>     <bch:tm> <bch:tm> <bch:byt>   <bch:tm> <int>
1 original     17.7s      18s    60.3MB      3.01m    10
2 update          3s    3.15s    56.5MB     31.38s    10
```

I'm sure further optimisations are possible - e.g. the phylogenies are already loaded by read.beast, and these could be directly passed into `read.stats_beast_internal`, to save the repeated text->phylo->text conversion there.